+++ Microting.TimePlanningBase/Infrastructure/Data/Entities/PlanRegistrationVersion.cs
@@ -158,6 +158,6 @@ public class PlanRegistrationVersion : PnBase
     public int PlannedEndOfShift5 { get; set; }
     public int PlannedBreakOfShift5 { get; set; }
     public bool PlanChangedByAdmin { get; set; }
-    public int NettoHoursOverride { get; set; }
+    public double NettoHoursOverride { get; set; }
     public bool NettoHoursOverrideActive { get; set; }
 }
